Update some comments
